fix(admin): trim standalone trace bloat and drop duplicate browser deps
Admin builds use outputFileTracingRoot at repo root but lacked LMS
outputFileTracingExcludes, so playwright/puppeteer and client-only libs
could land in apps/admin/.next/standalone. Share exclude globs via
scripts/next-standalone-trace-excludes.mjs, remove unused serverExternalPackages
(studio/cli, puppeteer, playwright, chromium), run prune-admin-standalone
after Northflank next build (keeps Remotion for lesson video API), and add
audit:admin-standalone to detect regressions.
