+++ AGENTS.md
@@ -854,7 +854,7 @@ Durable cannot CNAME-flatten apex to Northflank. **Do not** use apex **A** → N
 - **pnpm in Docker:** Store dir is **`/app/.pnpm-store`** (same filesystem as `/app/node_modules` for hardlinks). Never use `--package-import-method=copy` in Northflank Dockerfiles — it duplicates the full store into `node_modules` and triggers ENOSPC even when the registry cache mount works. Flow: `pnpm fetch` after copying workspace manifests, `COPY . .`, then `pnpm install --offline`.
 - **Build context:** `.dockerignore` excludes `node_modules`, `.next`, `supabase/`, `tests/`, `docs/`, etc. Verify with `tar -cf - --exclude-from=.dockerignore . | wc -c` (expect well under 1 GB).
 - **LMS:** service `elevate-lms`, Dockerfile `Dockerfile.northflank-lms`, branch `main`, readiness path `/api/ping`
-- **Admin (Northflank):** service `elevate-admin`, Dockerfile `Dockerfile.northflank-admin`, branch `main`, readiness path `/api/ping` — use `COPY . .` then cached `pnpm install --frozen-lockfile --package-import-method=copy` (workspace manifests must exist before install; avoid `--package-import-method=hardlink` on Northflank builders).
+- **Admin (Northflank):** service `elevate-admin`, Dockerfile `Dockerfile.northflank-admin`, branch `main`, readiness path `/api/ping`. Same pnpm fetch/offline flow as LMS (no `--package-import-method=copy`). After `next build`, **`node scripts/prune-admin-standalone.mjs`** drops playwright/puppeteer/three/etc. from `apps/admin/.next/standalone` but **keeps Remotion** for `/api/admin/generate-lesson-videos`. LMS uses `scripts/prune-standalone.mjs` via root `postbuild` and excludes Remotion via `outputFileTracingExcludes` (`scripts/next-standalone-trace-excludes.mjs`). Audit sizes: `node scripts/audit-admin-standalone-size.mjs` (fails if playwright/puppeteer remain).
 - **Admin (AWS ECS, legacy):** `elevate-admin-service` on `elevate-cluster`, `Dockerfile.admin`, CodeBuild `elevate-admin-build`. Deploy: `.github/workflows/deploy-admin.yml` on `main`.
 - `apps/admin/server.js` must load `.next/required-server-files.json` at startup.
 - **BuildKit cache:** `pnpm tsx scripts/northflank/ensure-build-cache.ts --execute` (10GB `useCache` on both services)
