+++ scripts/next-standalone-trace-excludes.mjs
@@ -1,0 +1,121 @@
+/**
+ * Shared outputFileTracingExcludes patterns for Next.js standalone builds.
+ *
+ * LMS (root next.config.mjs) and admin (apps/admin/next.config.mjs) both set
+ * outputFileTracingRoot to the repo root, so file tracing can pull the entire
+ * monorepo dependency graph into standalone unless these globs are applied.
+ */
+
+/** Safe on both LMS and admin — browser-only, dev tools, duplicate natives. */
+export const sharedStandaloneTraceExcludes = [
+  'public/generated/**',
+  'public/generated-images/**',
+  'reports/**',
+  'audit-packet/**',
+  'playwright-report/**',
+  // Browser automation (only used under scripts/, never admin/LMS API routes)
+  '**/node_modules/playwright/**',
+  '**/node_modules/puppeteer/**',
+  '**/node_modules/@playwright/**',
+  '**/node_modules/playwright-core/**',
+  '**/node_modules/puppeteer-core/**',
+  '**/node_modules/**/chromium/**',
+  '**/node_modules/@sparticuz/**',
+  '**/node_modules/chrome-aws-lambda/**',
+  '**/node_modules/chromium-bidi/**',
+  // Dev-only tools
+  '**/node_modules/typescript/**',
+  '**/node_modules/jsdom/**',
+  '**/node_modules/core-js/**',
+  // Sharp: platform binaries copied explicitly in Dockerfile.northflank-admin
+  '**/node_modules/@img/sharp-libvips-*/**',
+  '**/node_modules/@img/sharp-linux-*/**',
+  '**/node_modules/@img/sharp-darwin-*/**',
+  '**/node_modules/@img/sharp-win32-*/**',
+  '**/node_modules/jspdf/dist/**',
+  '**/node_modules/monaco-editor/**',
+  '**/node_modules/.pnpm/monaco-editor*/**',
+  '**/node_modules/@monaco-editor/**',
+  '**/node_modules/.pnpm/@monaco-editor*/**',
+  '**/node_modules/node-pty/**',
+  '**/node_modules/.pnpm/node-pty*/**',
+  '**/node_modules/lucide-react/**',
+  '**/node_modules/.pnpm/lucide-react*/**',
+  '**/node_modules/three/**',
+  '**/node_modules/.pnpm/three*/**',
+  '**/node_modules/three-stdlib/**',
+  '**/node_modules/.pnpm/three-stdlib*/**',
+  '**/node_modules/@react-three/**',
+  '**/node_modules/.pnpm/@react-three*/**',
+  '**/node_modules/@dimforge/**',
+  '**/node_modules/.pnpm/@dimforge*/**',
+  '**/node_modules/jspdf/**',
+  '**/node_modules/.pnpm/jspdf*/**',
+  '**/node_modules/pdfjs-dist/**',
+  '**/node_modules/.pnpm/pdfjs-dist*/**',
+  '**/node_modules/hls.js/**',
+  '**/node_modules/.pnpm/hls.js*/**',
+  '**/node_modules/canvas/**',
+  '**/node_modules/.pnpm/canvas*/**',
+  '**/node_modules/@napi-rs/**',
+  '**/node_modules/.pnpm/@napi-rs*/**',
+  '**/node_modules/@mediapipe/**',
+  '**/node_modules/.pnpm/@mediapipe*/**',
+  '**/node_modules/video.js/**',
+  '**/node_modules/.pnpm/video.js*/**',
+  '**/node_modules/@videojs/**',
+  '**/node_modules/.pnpm/@videojs*/**',
+  '**/node_modules/@sentry/cli/**',
+  '**/node_modules/.pnpm/@sentry+cli*/**',
+  '**/node_modules/prettier/**',
+  '**/node_modules/.pnpm/prettier*/**',
+  '**/node_modules/es-toolkit/**',
+  '**/node_modules/.pnpm/es-toolkit*/**',
+  '**/node_modules/recharts/**',
+  '**/node_modules/.pnpm/recharts*/**',
+  '**/node_modules/mediabunny/**',
+  '**/node_modules/.pnpm/mediabunny*/**',
+  '**/node_modules/fontkit/**',
+  '**/node_modules/.pnpm/fontkit*/**',
+  '**/node_modules/hyphen/**',
+  '**/node_modules/.pnpm/hyphen*/**',
+  '**/node_modules/web-streams-polyfill/**',
+  '**/node_modules/.pnpm/web-streams-polyfill*/**',
+  '**/node_modules/@swc/core/**',
+  '**/node_modules/.pnpm/@swc+core*/**',
+  '**/node_modules/happy-dom/**',
+  '**/node_modules/.pnpm/happy-dom*/**',
+  // TypeScript sources — compiled output lives under .next/server
+  'app/**/*.tsx',
+  'app/**/*.ts',
+  'components/**/*.tsx',
+  'components/**/*.ts',
+  'lib/**/*.ts',
+  'lib/**/*.tsx',
+];
+
+/**
+ * LMS-only: Remotion/TTS/ffmpeg and pdf-lib are not needed in the main SSR bundle.
+ * Video generation runs on admin / Railway workers.
+ */
+export const lmsOnlyStandaloneTraceExcludes = [
+  '**/node_modules/pdf-lib/**',
+  '**/node_modules/.pnpm/pdf-lib*/**',
+  '**/node_modules/@remotion/**',
+  '**/node_modules/.pnpm/@remotion*/**',
+  '**/node_modules/remotion/**',
+  '**/node_modules/.pnpm/remotion*/**',
+  '**/node_modules/edge-tts/**',
+  '**/node_modules/.pnpm/edge-tts*/**',
+  '**/node_modules/ffmpeg-static/**',
+  '**/node_modules/.pnpm/ffmpeg-static*/**',
+  '**/node_modules/@ffmpeg/**',
+  '**/node_modules/.pnpm/@ffmpeg*/**',
+  'remotion-src/**',
+  '**/node_modules/@rspack/**',
+  '**/node_modules/.pnpm/@rspack*/**',
+  '**/node_modules/pdf-parse/**',
+  '**/node_modules/.pnpm/pdf-parse*/**',
+  '**/node_modules/pdfkit/**',
+  '**/node_modules/.pnpm/pdfkit*/**',
+];
